Import add date from Delicious export
This adds an $added parameter to
OCA\\Bookmarks\\Controller\\Lib\\Bookmarks::addBookmark(), which
defaults to 0, in which case the current time (using UNIX_TIMESTAMP() in
SQL) is used. If positive, this is used as the time of addition (and
last modification) of the link.

When modifying an existing link, $added is ignored, and only the last
modification date is updated to UNIX_TIMESTAMP().

Signed-off-by: Olivier Mehani <shtrom@ssji.net>