Update error handling on "Import"
The test for `response.ok` doesn't appear to work in latest Brave or Chrome.  The result is that, although the POST to `/import` succeeded, it throws an error with `[Object object] [Object object] ...`.  This test ensures that errors are thrown only when `response.data.success` is not set correctly.  The output of `response.data.data.join('\n')` was worthless in the case of my bug so I wasn't sure which style should be used on actual error scenarios.  Please advise.

Signed-off-by: Andy Jacobs <andy@seattledevelopers.coop>